[feat] add writtenContent Component
