+++ src/components/reput/index.ts
@@ -1,1 +1,2 @@
 export * from './myreput';
+export * from './writtenreput';
