+++ src/components/reput/writtenreput/index.ts
@@ -1,0 +1,1 @@
+export * from './WrittenContent';
