+++ src/styles/components/reput/writtenContent.styles.ts
